# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerDelta": {
        "properties": {
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "seq": {
            "default": 0,
            "title": "Seq",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "ts": {
            "format": "date-time",
            "title": "Ts",
            "type": "string"
          },
          "type": {
            "const": "answer.delta",
            "default": "answer.delta",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AnswerDelta",
        "type": "object"
      },
...
      "Body_upload_library_document_library_documents_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_library_document_library_documents_post",
        "type": "object"
      },
...
      "ChunkOut": {
        "properties": {
          "heading_path": {
            "title": "Heading Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "ord": {
            "title": "Ord",
            "type": "integer"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "section_id",
          "ord",
          "page",
          "heading_path",
          "text",
          "metadata"
        ],
        "title": "ChunkOut",
        "type": "object"
      },
...
      "DocumentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_flags": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PageFlag"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Flags"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "mime",
          "sha256",
          "status",
          "page_flags",
          "error",
          "tags",
          "created_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
      "DocumentPatch": {
        "properties": {
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "tags"
        ],
        "title": "DocumentPatch",
        "type": "object"
      },
      "DocumentUploadOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deduped": {
            "default": false,
            "title": "Deduped",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_flags": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PageFlag"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Flags"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "mime",
          "sha256",
          "status",
          "page_flags",
          "error",
          "tags",
          "created_at"
        ],
        "title": "DocumentUploadOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LibraryDocumentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "editable": {
            "title": "Editable",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "page_flags": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PageFlag"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Flags"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "shared": {
            "title": "Shared",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "mime",
          "sha256",
          "status",
          "page_flags",
          "error",
          "tags",
          "created_at",
          "shared",
          "editable"
        ],
        "title": "LibraryDocumentOut",
        "type": "object"
      },
      "LibraryOut": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/LibraryDocumentOut"
            },
            "title": "Documents",
            "type": "array"
          },
          "starter_questions": {
            "items": {
              "type": "string"
            },
            "title": "Starter Questions",
            "type": "array"
          }
        },
        "required": [
          "documents",
          "starter_questions"
        ],
        "title": "LibraryOut",
        "type": "object"
      },
...
      "PageFlag": {
        "properties": {
          "flags": {
            "items": {
              "type": "string"
            },
            "title": "Flags",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "flags"
        ],
        "title": "PageFlag",
        "type": "object"
      },
      "PinRequest": {
        "properties": {
          "target": {
            "default": "chat",
            "pattern": "^chat$",
            "title": "Target",
            "type": "string"
          }
        },
        "title": "PinRequest",
        "type": "object"
      },
      "PinResponse": {
        "properties": {
          "document_id": {
            "title": "Document Id",
            "type": "string"
          }
        },
        "required": [
          "document_id"
        ],
        "title": "PinResponse",
        "type": "object"
      },
...
      "RunCancelResponse": {
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunCancelResponse",
        "type": "object"
      },
...
      "RunStreamResponse": {
        "description": "References the union so it lands in OpenAPI (TRD \u00a712: the TS union is\ngenerated from the schema, never hand-rolled).",
        "discriminator": {
          "mapping": {
            "abstain": "#/components/schemas/Abstain",
            "answer.delta": "#/components/schemas/AnswerDelta",
            "answer.hold": "#/components/schemas/AnswerHold",
            "conflict": "#/components/schemas/Conflict",
            "decision": "#/components/schemas/Decision",
            "heartbeat": "#/components/schemas/Heartbeat",
            "metrics": "#/components/schemas/Metrics",
            "plan": "#/components/schemas/Plan",
            "retrieval": "#/components/schemas/Retrieval",
            "review.claim": "#/components/schemas/ReviewClaim",
            "revision": "#/components/schemas/Revision",
            "run.cancelled": "#/components/schemas/RunCancelled",
            "run.completed": "#/components/schemas/RunCompleted",
            "run.failed": "#/components/schemas/RunFailed",
            "run.started": "#/components/schemas/RunStarted",
            "step.completed": "#/components/schemas/StepCompleted",
            "step.started": "#/components/schemas/StepStarted",
            "suggestions": "#/components/schemas/Suggestions",
            "thinking.delta": "#/components/schemas/ThinkingDelta"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/RunStarted"
          },
          {
            "$ref": "#/components/schemas/StepStarted"
          },
          {
            "$ref": "#/components/schemas/StepCompleted"
          },
          {
            "$ref": "#/components/schemas/Decision"
          },
          {
            "$ref": "#/components/schemas/ThinkingDelta"
          },
          {
            "$ref": "#/components/schemas/Plan"
          },
          {
            "$ref": "#/components/schemas/Retrieval"
          },
          {
            "$ref": "#/components/schemas/AnswerDelta"
          },
          {
            "$ref": "#/components/schemas/Abstain"
          },
          {
            "$ref": "#/components/schemas/Conflict"
          },
          {
            "$ref": "#/components/schemas/ReviewClaim"
          },
          {
            "$ref": "#/components/schemas/AnswerHold"
          },
          {
            "$ref": "#/components/schemas/Revision"
          },
          {
            "$ref": "#/components/schemas/Suggestions"
          },
          {
            "$ref": "#/components/schemas/Metrics"
          },
          {
            "$ref": "#/components/schemas/Heartbeat"
          },
          {
            "$ref": "#/components/schemas/RunCompleted"
          },
          {
            "$ref": "#/components/schemas/RunCancelled"
          },
          {
            "$ref": "#/components/schemas/RunFailed"
          }
        ],
        "title": "RunStreamResponse"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Veriforge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "library"
        ]
      },
      "get": {
        "operationId": "get_document_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "library"
        ]
      },
      "patch": {
        "operationId": "patch_document_documents__document_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Document",
        "tags": [
          "library"
        ]
      }
    },
    "/documents/{document_id}/chunks": {
      "get": {
        "operationId": "get_document_chunks_documents__document_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChunkOut"
                  },
                  "title": "Response Get Document Chunks Documents  Document Id  Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document Chunks",
        "tags": [
          "library"
        ]
      }
    },
    "/documents/{document_id}/reindex": {
      "post": {
        "operationId": "reindex_document_documents__document_id__reindex_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reindex Document",
        "tags": [
          "library"
        ]
      }
    },
...
    "/library": {
      "get": {
        "operationId": "get_library_library_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Library",
        "tags": [
          "library"
        ]
      }
    },
    "/library/documents": {
      "post": {
        "operationId": "upload_library_document_library_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_library_document_library_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Library Document",
        "tags": [
          "library"
        ]
      }
    },
...
    "/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/stream": {
      "get": {
        "operationId": "stream_run_runs__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Seq",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStreamResponse"
                }
              }
            },
            "description": "SSE event stream"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stream Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/web-sources/{web_page_id}/pin": {
      "post": {
        "operationId": "pin_web_source_web_sources__web_page_id__pin_post",
        "parameters": [
          {
            "in": "path",
            "name": "web_page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Web Page Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pin Web Source",
        "tags": [
          "web-sources"
        ]
      }
    }
  }
}